Refactor MdnsServices initialization and property access
This commit refactors the `MdnsServices.kt` file by moving the initialization of `hosts` and `services` to their declaration sites, removing the `init` block.

Additionally, it replaces calls to the `getServiceInfo()` method with direct access to the `serviceInfo` property and uses an underscore `_` for the unused exception variable in a `catch` block.